Various fixes for stuff found in testing.
 * Some unnecessary logging removed
 * Use MariaDB by default
 * Create tmp directory if not there
 * Don't error if tracking server versions when options page is build
 * Diverging versions of JavaFX (probably main problem)

+++ client-logonbox-vpn-gui-jfx/src/main/java/com/logonbox/vpn/client/gui/jfx/UI.java
@@ -3,6 +3,7 @@ package com.logonbox.vpn.client.gui.jfx;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
+import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.io.PrintWriter;
@@ -189,14 +190,14 @@ public class UI extends AbstractController implements Listener {
 	}
 
 	final static ResourceBundle bundle = ResourceBundle.getBundle(UI.class.getName());
-	
+
 	static {
 		ToasterSettings settings = new ToasterSettings();
 		settings.setAppName(bundle.getString("appName"));
 		settings.setSystemTrayIconMode(SystemTrayIconMode.HIDDEN);
 		ToasterFactory.setSettings(settings);
 	}
-	
+
 	static int DROP_SHADOW_SIZE = 11;
 
 	final static Logger log = LoggerFactory.getLogger(UI.class);
@@ -259,15 +260,14 @@ public class UI extends AbstractController implements Listener {
 	public void bridgeEstablished() {
 
 		LOG.info("Bridge established.");
-		
+
 		/*
 		 * If we were waiting for this, it's part of the update process. We don't want
 		 * the connection continuing
 		 */
 		if (awaitingBridgeEstablish != null) {
 			awaitingRestart = true;
-		}
-		else {
+		} else {
 			reloadState();
 		}
 
@@ -1028,8 +1028,7 @@ public class UI extends AbstractController implements Listener {
 					 * couldn't get a custom URL handler to work either).
 					 */
 					byte[] localCss = IOUtils.toByteArray(context.getCustomLocalWebCSSFile().toURI());
-					String datauri = "data:text/css;base64," + DatatypeConverter
-							.printBase64Binary(localCss);
+					String datauri = "data:text/css;base64," + DatatypeConverter.printBase64Binary(localCss);
 					webView.getEngine().setUserStyleSheetLocation(datauri);
 
 					setupPage();
@@ -1259,26 +1258,28 @@ public class UI extends AbstractController implements Listener {
 			branding = context.getBridge().getClientService().getBranding(getSelectedConnection());
 			if (branding == null) {
 				log.info(String.format("Removing branding."));
-				if(logoFile != null) {
+				if (logoFile != null) {
 					logoFile.delete();
 				}
-			}
-			else {
+			} else {
 				log.info(String.format("Adding custom branding"));
 				String logo = branding.getLogo();
-				if(StringUtils.isNotBlank(logo)) {
+				if (StringUtils.isNotBlank(logo)) {
 					log.info(String.format("Attempting to cache logo"));
 					String basename = FilenameUtils.getExtension(logo);
-					if(!basename.equals(""))
-						basename= "." + basename;
-					File newLogoFile = new File("tmp" + File.separator + "logo" + basename);
-					if(logoFile != null && !newLogoFile.equals(logoFile)) {
-						logoFile.delete();
-					}
+					if (!basename.equals(""))
+						basename = "." + basename;
+					File tmpDir = new File("tmp");
 					try {
+						if (!tmpDir.exists() && !tmpDir.mkdirs())
+							throw new IOException(String.format("Failed to create temporary directory %s.", tmpDir));
+						File newLogoFile = new File(tmpDir, "logo" + basename);
+						if (logoFile != null && !newLogoFile.equals(logoFile)) {
+							logoFile.delete();
+						}
 						URL logoUrl = new URL(logo);
-						try(InputStream urlIn = logoUrl.openStream()) {
-							try(OutputStream out = new FileOutputStream(newLogoFile)) {
+						try (InputStream urlIn = logoUrl.openStream()) {
+							try (OutputStream out = new FileOutputStream(newLogoFile)) {
 								urlIn.transferTo(out);
 							}
 						}
@@ -1286,14 +1287,13 @@ public class UI extends AbstractController implements Listener {
 						logoFile = newLogoFile;
 						branding.setLogo(newLogoFile.toURI().toString());
 						log.info(String.format("Logo cached from %s to %s", logoUrl, newLogoFile.toURI()));
-					}
-					catch(Exception e) {
+					} catch (Exception e) {
 						log.error(String.format("Failed to cache logo"), e);
 					}
 				}
 			}
 		} catch (RemoteException e) {
-			throw new IllegalStateException("Impossible!", e);
+			throw new IllegalStateException("Failed to retrieve state from local service.", e);
 		}
 	}
